Fix hash table entry iteration
Pointer subtraction to get the start of the buckets array, as done by
the TBL_FOR_EACH macro, caused undefined behavior on empty tables,
since the buckets have yet to be allocated. Replace it with a function
that consults the control bytes before indexing into buckets.

This way, the type of the keys does not need to be stored. Therefore a
single struct Table type suffices, meaning find_insert_table, etc. may
be generic for all tables.